requirements.txt
# =============================================================================
# DS19 — Week 6 Dependencies (All Conflicts Resolved)
# Python Version: 3.9 or 3.10 (Recommended for Windows)
# =============================================================================

# ─────────────────────────────────────────────────────────────
# BUILD TOOLS (Install these FIRST to avoid conflicts)
# ─────────────────────────────────────────────────────────────
packaging>=24.0,<25.0
protobuf>=5.0,<7.0
wheel>=0.46.0
setuptools>=68.0.0

# ─────────────────────────────────────────────────────────────
# CORE DATA
# ─────────────────────────────────────────────────────────────
numpy
pandas
scipy==1.11.4
pyarrow==14.0.1

# ─────────────────────────────────────────────────────────────
# MACHINE LEARNING
# ─────────────────────────────────────────────────────────────
torch==2.1.2
scikit-learn==1.3.2
lightgbm==4.1.0
faiss-cpu==1.7.4

# ─────────────────────────────────────────────────────────────
# FEATURE STORE
# ─────────────────────────────────────────────────────────────
feast

# ─────────────────────────────────────────────────────────────
# MLOPS & MONITORING
# ─────────────────────────────────────────────────────────────
mlflow>=2.9.0,<3.0.0
prometheus-client==0.18.0

# ─────────────────────────────────────────────────────────────
# BACKEND API
# ─────────────────────────────────────────────────────────────
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
python-dotenv==1.0.0

# ─────────────────────────────────────────────────────────────
# UTILITIES & VISUALIZATION
# ─────────────────────────────────────────────────────────────
loguru==0.7.2
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0